refactored axios syntax to make post request instead of using curl
